Make sources same across 5.6 and 5.7 to allow easier patching

+++ 5.6/root/usr/share/container-scripts/mysql/passwd-change.sh
@@ -9,17 +9,29 @@ fi
 # The MYSQL_ROOT_PASSWORD is optional, therefore we need to either enable remote
 # access with a password if the variable is set or disable remote access otherwise.
 if [ -v MYSQL_ROOT_PASSWORD ]; then
-  # GRANT will create a user if it doesn't exist and set its password
+  # create a user if it doesn't exist and set its password
+  # for 5.6 and lower we use the trick that GRANT creates a user if not exists
+  # because IF NOT EXISTS clause does not exist in that versions yet
+  if [[ "$MYSQL_VERSION" > "5.6" ]] ; then
+    mysql $mysql_flags <<EOSQL
+      CREATE USER IF NOT EXISTS 'root'@'%';
+EOSQL
+  fi
   mysql $mysql_flags <<EOSQL
     GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY '${MYSQL_ROOT_PASSWORD}' WITH GRANT OPTION;
 EOSQL
 else
-  # We do GRANT and DROP USER to emulate a DROP USER IF EXISTS statement
-  # http://bugs.mysql.com/bug.php?id=19166
-  mysql $mysql_flags <<EOSQL
-    GRANT USAGE ON *.* TO 'root'@'%';
-    DROP USER 'root'@'%';
-    FLUSH PRIVILEGES;
+  if [[ "$MYSQL_VERSION" < "5.7" ]] ; then
+    mysql $mysql_flags <<EOSQL
+      GRANT USAGE ON *.* TO 'root'@'%';
+      DROP USER 'root'@'%';
+      FLUSH PRIVILEGES;
 EOSQL
+  else
+    mysql $mysql_flags <<EOSQL
+      DROP USER IF EXISTS 'root'@'%';
+      FLUSH PRIVILEGES;
+EOSQL
+  fi
 fi
